Make debug helpers test only
