+++ lib/hypothesis.rb
@@ -4,7 +4,6 @@ require 'hypothesis/errors'
 require 'hypothesis/providers'
 require 'hypothesis/engine'
 require 'hypothesis/world'
-require 'hypothesis/debug'
 
 module Hypothesis
   HYPOTHESIS_LOCATION = File.dirname(__FILE__)
